fix a bug

+++ install.sh
@@ -9,11 +9,17 @@ for file in $files; do
 	echo "--------------------"
 	if [ -f ~/".$file" ]; then
 		echo "Config .$file already exists!"
-		echo "Backup :   .$file ---> .$file.bck"
-		mv ~/".$file" ~/".$file.bck"
-		ln -s "$loc/$file" ~/".$file"
-		if [ $? -eq 0 ]; then
-			echo "Success!"
+
+		# If there is a backup, we do nothing
+		if [ -f ~/".$file.bck" ]; then
+			echo "If you want to install $file, Please rename .$file.bck"
+		else
+			echo "Backup :   .$file ---> .$file.bck"
+			mv ~/".$file" ~/".$file.bck"
+			ln -s "$loc/$file" ~/".$file"
+			if [ $? -eq 0 ]; then
+				echo "Success!"
+			fi
 		fi
 	else
 		ln -s "$loc/$file" ~/".$file"
